+++ Final-project/src/main/java/com/kh/finalproject/experience/model/service/ExperienceServiceImpl.java
@@ -70,13 +70,15 @@ public class ExperienceServiceImpl implements ExperienceService {
 		int result = experienceDao.updateExperience(sqlSession, exp);
 		System.out.println(result);
 		System.out.println("게시글 업데이트 하나");
-		for(String oldFile : oldFiles) {
-			System.out.println("비어있니?");
-			System.out.println(oldFile.isEmpty());
-			System.out.println(oldFile.equals(""));
-			if(!oldFile.isEmpty()) result *= experienceDao.deleteFiles(sqlSession, oldFile);
-			System.out.println(result);
-			System.out.println("문제찾기 파일 딜리트");
+		if(oldFiles != null) {
+			for(String oldFile : oldFiles) {
+				System.out.println("비어있니?");
+				System.out.println(oldFile.isEmpty());
+				System.out.println(oldFile.equals(""));
+				if(!oldFile.isEmpty()) result *= experienceDao.deleteFiles(sqlSession, oldFile);
+				System.out.println(result);
+				System.out.println("문제찾기 파일 딜리트");
+			}
 		}
 		for(Attachment file : fileList) {
 			result *= commonDao.insertFiles(sqlSession, file);
